add: prepare for local DB -> Postgres
