+++ .dev.env
@@ -1,0 +1,1 @@
+DB_PASSWORD=8579ygh
