+++ src/app.module.ts
@@ -9,11 +9,11 @@ import { typeORMConfig } from './config/typeorm.config';
   imports: [
     BoardsModule,
     TypeOrmModule.forRoot(typeORMConfig),
-    ConfigModule.forRoot({
-      isGlobal: true,
-      envFilePath: process.env.NODE_ENV === 'dev' ? '.dev.env' : '.test.env',
-      ignoreEnvFile: process.env.NODE_ENV === 'prod',
-  }),
+  //   ConfigModule.forRoot({
+  //     isGlobal: true,
+  //     envFilePath: process.env.NODE_ENV === 'dev' ? '.dev.env' : '.test.env',
+  //     ignoreEnvFile: process.env.NODE_ENV === 'prod',
+  // }),
   ],
 })
 export class AppModule {}
