--- src/boards/boards.model.ts
@@ -1,12 +1,0 @@
-// have to connect this with "boards.service"
-export interface Board {
-    id: string;
-    title: string;
-    description: string;
-    status: BoardStatus;
-}
-
-export enum BoardStatus {
-    PUBLIC = 'PUBLIC',
-    PRIVATE = 'PRIVATE',
-}
