+++ src/boards/boards.module.ts
@@ -1,8 +1,14 @@
 import { Module } from '@nestjs/common';
-import { BoardsController } from './boards.controller';
+import { TypeOrmModule } from '@nestjs/typeorm';
+
 import { BoardsService } from './boards.service';
+import { BoardRepository } from './board.repository';
+import { BoardsController } from './boards.controller';
 
 @Module({
+  imports: [
+    TypeOrmModule.forFeature([BoardRepository])
+  ],
   controllers: [BoardsController],
   providers: [BoardsService]
 })
